+++ src/constants.js
@@ -32,6 +32,7 @@ export const requestStatuses = {
 
 export const itemStatuses = {
   PAGED: 'Paged',
+  DECLARED_LOST: 'Declared lost',
 };
 
 export const requestTypesMap = {